+++ lib/ts/ingredients/smsdelivery/services/supertokens.ts
@@ -13,8 +13,4 @@
  * under the License.
  */
 
-export type SupertokensServiceConfig = {
-    apiKey: string;
-};
-
 export const SUPERTOKENS_SMS_SERVICE_URL = "https://api.supertokens.com/0/services/sms";